common: Add ManagedIPAMAddCall function
This function wraps all the steps that the CmdAdd function took around
the actual ipam.Add call with a function. It also brings some other
private dependencies into common.
